move the mg related commands to a separate package
